moved UpdateArticle class to the models folder
