+++ src/Forem.Api/ArticlesService.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Forem.Api.Internal;
 using Forem.Api.Models;
 
 namespace Forem.Api
